refactor(providers): extract the form-mapping layer to lib/providerForm
ProvidersPanel opened with ~200 lines of pure mapping helpers
(FormState/FormModel shapes, toFormEfforts, asFormModel, emptyForm,
modelsFromProvider, effortsFromProvider, formFromPreset, hostOf,
ccSwitchStatusKey) living above the component with no tests.

Move them to src/lib/providerForm with 10 unit tests covering the
mapping edges (absent/empty effort lists, legacy single-model
fallback, preset endpoint application, host parsing, cc-switch
status fallback). The panel keeps only the presentational FieldHelp.
Pure move; no behavior change.
