+++ src/components/ProvidersPanel.tsx
@@ -12,6 +12,19 @@ import {
   type MouseEvent,
 } from "react";
 import * as api from "@/lib/api";
+import {
+  asFormModel,
+  toFormEfforts,
+  ccSwitchStatusKey,
+  emptyForm,
+  effortsFromProvider,
+  formFromPreset,
+  hostOf,
+  modelsFromProvider,
+  type FormState,
+  type RightMode,
+  type Selection,
+} from "@/lib/providerForm";
 import { createT, type Locale, type MessageKey } from "@/i18n";
 import { Select } from "@/components/Select";
 import { GlassModal } from "@/components/GlassModal";
@@ -52,7 +65,6 @@ import {
 } from "@/lib/providerRouteHonesty";
 import {
   PROVIDER_PRESETS,
-  alignGrokPresetEfforts,
   applyPresetEndpoint,
   defaultCustomChannelEfforts,
   matchPresetEndpoint,
@@ -95,45 +107,6 @@ export interface ProvidersPanelProps {
   ) => void;
 }
 
-type FormEffort = {
-  id: string;
-  name: string;
-  isDefault: boolean;
-};
-
-type FormModel = {
-  /** Upstream request body model id. */
-  id: string;
-  /** Display name shown on composer chip / menu. */
-  name: string;
-  contextWindow?: number | null;
-  supportsVision?: boolean;
-  supportsVideo?: boolean;
-  efforts?: FormEffort[];
-};
-
-function toFormEfforts(
-  list?: Array<{ id: string; name?: string; isDefault?: boolean }>,
-): FormEffort[] | undefined {
-  if (!list?.length) return list ? [] : undefined;
-  return list.map((e) => ({
-    id: e.id,
-    name: e.name?.trim() || e.id,
-    isDefault: !!e.isDefault,
-  }));
-}
-
-function asFormModel(m: api.ProviderModelEntry): FormModel {
-  return {
-    id: m.id,
-    name: m.name,
-    contextWindow: m.contextWindow ?? null,
-    supportsVision: m.supportsVision,
-    supportsVideo: m.supportsVideo,
-    efforts: toFormEfforts(m.efforts),
-  };
-}
-
 function FieldHelp({ label, tip }: { label: string; tip: string }) {
   return (
     <span className="prov-field__label">
@@ -155,162 +128,6 @@ function FieldHelp({ label, tip }: { label: string; tip: string }) {
       ) : null}
     </span>
   );
-}
-
-type FormState = {
-  id: string;
-  name: string;
-  baseUrl: string;
-  /** When true, host keeps baseUrl as-is (no auto `/v1`). */
-  baseUrlFullPath: boolean;
-  apiKey: string;
-  apiBackend: string;
-  providerMode: "generic" | "grok_build_proxy";
-  models: FormModel[];
-  efforts: FormEffort[];
-  /** Extra rules appended to the system prompt on this channel. */
-  appendPrompt: string;
-  /** Explicit: this relay accepts image pixels. */
-  supportsVision: boolean;
-  /** External signup URL for “Get API Key” (from preset). */
-  apiKeyUrl: string | null;
-  extraHeaders: { name: string; value: string }[];
-  /** Prefill / keep per-channel context_window. Null = omit on create. */
-  contextWindow: number | null;
-};
-
-type RightMode = "empty" | "pick" | "create" | "edit" | "official";
-type Selection = null | "official" | string;
-
-const emptyForm = (): FormState => ({
-  id: "",
-  name: "",
-  baseUrl: "",
-  baseUrlFullPath: false,
-  apiKey: "",
-  apiBackend: "responses",
-  providerMode: "generic",
-  appendPrompt: "",
-  supportsVision: false,
-  models: [],
-  efforts: defaultCustomChannelEfforts().map((e) => ({
-    id: e.id,
-    name: e.name || e.id,
-    isDefault: !!e.isDefault,
-  })),
-  apiKeyUrl: null,
-  extraHeaders: [],
-  contextWindow: null,
-});
-
-function modelsFromProvider(p: api.CustomProvider): FormModel[] {
-  if (p.models?.length) {
-    return p.models.map((m) => ({
-      id: m.id,
-      name: m.name?.trim() || m.id,
-      contextWindow: m.contextWindow ?? null,
-      supportsVision: m.supportsVision,
-      supportsVideo: m.supportsVideo,
-      efforts: m.efforts?.map((e) => ({
-        id: e.id,
-        name: e.name?.trim() || e.id,
-        isDefault: !!e.isDefault,
-      })),
-    }));
-  }
-  const id = p.model?.trim() ?? "";
-  if (!id) return [];
-  return [{ id, name: id }];
-}
-
-function effortsFromProvider(p: api.CustomProvider): FormEffort[] {
-  const aligned = alignGrokPresetEfforts({
-    providerId: p.id,
-    baseUrl: p.baseUrl,
-    efforts: p.efforts,
-  });
-  const source = aligned ?? p.efforts;
-  if (source?.length) {
-    return source.map((e) => ({
-      id: e.id,
-      name: e.name?.trim() || e.id,
-      isDefault: !!e.isDefault,
-    }));
-  }
-  return defaultCustomChannelEfforts().map((e) => ({
-    id: e.id,
-    name: e.name || e.id,
-    isDefault: !!e.isDefault,
-  }));
-}
-
-function formFromPreset(preset: ProviderPreset, endpointId?: string): FormState {
-  const applied = applyPresetEndpoint(
-    preset,
-    endpointId ?? preset.defaultEndpointId,
-  );
-  return {
-    id: preset.suggestedId,
-    name: preset.name,
-    baseUrl: applied.baseUrl,
-    // Most presets ship with `/v1`; Volcengine Ark / Zhipu roots are already complete.
-    baseUrlFullPath: applied.baseUrlFullPath,
-    apiKey: "",
-    apiBackend: preset.apiBackend,
-    providerMode: "generic",
-    // Presets carry no channel rules — opt-in per provider.
-    appendPrompt: "",
-    supportsVision: !!preset.supportsVision,
-    models: preset.models.map((m) => ({
-      id: m.id,
-      name: m.name || m.id,
-      contextWindow: m.contextWindow ?? preset.contextWindow ?? null,
-      supportsVision: m.supportsVision ?? preset.supportsVision,
-      supportsVideo: m.supportsVideo,
-      efforts: (m.efforts?.length ? m.efforts : preset.efforts).map((e) => ({
-        id: e.id,
-        name: e.name || e.id,
-        isDefault: !!e.isDefault,
-      })),
-    })),
-    efforts: preset.efforts.map((e) => ({
-      id: e.id,
-      name: e.name || e.id,
-      isDefault: !!e.isDefault,
-    })),
-    apiKeyUrl: applied.apiKeyUrl,
-    extraHeaders: [],
-    contextWindow:
-      preset.contextWindow && preset.contextWindow > 0
-        ? preset.contextWindow
-        : null,
-  };
-}
-
-function hostOf(url: string): string {
-  try {
-    return new URL(url).host || url;
-  } catch {
-    return url;
-  }
-}
-
-function ccSwitchStatusKey(status: string): MessageKey {
-  switch (status) {
-    case "importable":
-      return "prov.ccSwitch.status.importable";
-    case "official":
-      return "prov.ccSwitch.status.official";
-    case "missing_key":
-      return "prov.ccSwitch.status.missing_key";
-    case "proxy_managed":
-      return "prov.ccSwitch.status.proxy_managed";
-    case "exists":
-      return "prov.ccSwitch.status.exists";
-    case "invalid":
-    default:
-      return "prov.ccSwitch.status.invalid";
-  }
 }
 
 export function ProvidersPanel({
